Revert "rework operator exec button to allow prompt"
This reverts commit 5a2c696ae3873639817df4a80c422267af10a195.
